Refactor to align domain and model package layout, included openapi UI within API.
