+++ quora-api/src/main/kotlin/com/bluurr/quora/config/QuoraClientConfiguration.kt
@@ -1,6 +1,6 @@
 package com.bluurr.quora.config
 
-import com.bluurr.quora.domain.user.LoginCredential
+import com.bluurr.quora.model.user.LoginCredential
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
