+++ quora-loader/src/main/java/com/bluurr/quora/model/user/InvalidLoginException.java
@@ -1,4 +1,4 @@
-package com.bluurr.quora.domain.user;
+package com.bluurr.quora.model.user;
 
 public class InvalidLoginException extends IllegalStateException {
   public InvalidLoginException(final String message) {
