+++ quora-loader/src/main/java/com/bluurr/quora/navigator/SearchPageNavigator.java
@@ -1,6 +1,6 @@
 package com.bluurr.quora.navigator;
 
-import com.bluurr.quora.domain.QuestionSearchResult;
+import com.bluurr.quora.model.QuestionSearchResult;
 import com.bluurr.quora.extension.EnhancedDriver;
 import com.bluurr.quora.page.component.InfiniteScrollIterator;
 import com.bluurr.quora.page.search.SearchPage;
