+++ quora-loader/src/main/java/com/bluurr/quora/page/search/SearchPage.java
@@ -1,6 +1,6 @@
 package com.bluurr.quora.page.search;
 
-import com.bluurr.quora.domain.QuestionSearchResult;
+import com.bluurr.quora.model.QuestionSearchResult;
 import com.bluurr.quora.extension.EnhancedDriver;
 import com.bluurr.quora.page.PageObject;
 import com.bluurr.quora.page.component.InfiniteScrollPage;
